Correcao Bugs / Classes DTOs

+++ src/main/java/com/eds/edstest/repositories/VeiculoRepository.java
@@ -3,5 +3,5 @@ package com.eds.edstest.repositories;
 import com.eds.edstest.entities.Veiculos;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface VeiculoRepository extends JpaRepository<Veiculos, Integer> {
+public interface VeiculoRepository extends JpaRepository<Veiculos, Long> {
 }
